Refactoring using some vue conventions
