+++ components/ArticleFeaturedImage.vue
