+++ pages/pages/_page.vue
@@ -1,6 +1,6 @@
 <template>
   <article class="single-article">
-    <FeaturedImage v-if="featuredImage" :expanded="expanded" :featuredImage="featuredImage"></FeaturedImage>
+    <article-featured-image v-if="featuredImage" :expanded="expanded" :featuredImage="featuredImage" />
     <transition name="slide-fade">
       <div class="narrow" :class="{ 'expanded': expanded, 'no-featured-image': !featuredImage }">
         <button class="expand-featured-image" title="Show full image" @click.prevent="expanded = !expanded" :class="{ 'expanded': expanded }" v-if="featuredImage.source_url">
@@ -25,7 +25,7 @@
 
 <script>
 import * as Vibrant from 'node-vibrant'
-import FeaturedImage from '~/components/FeaturedImage.vue'
+import ArticleFeaturedImage from '~/components/ArticleFeaturedImage.vue'
 
 if (process.browser) {
   require('lightgallery.js')
@@ -52,7 +52,7 @@ export default {
   },
 
   components: {
-    FeaturedImage
+    ArticleFeaturedImage
   },
 
   computed: {
